feat: move show and repay pages
