--- src/pages/[marketplace]/deals/[did]/show.tsx
@@ -1,30 +1,0 @@
-import React, { ReactElement } from "react";
-import Layout from "@components/Layout";
-import { NextPageWithLayout } from "pages/_app";
-import { getMarketDealsPaths } from "@utils/export.utils";
-import ShowDeal from "@components/ShowDeal";
-
-const Show: NextPageWithLayout = () => {
-	return <ShowDeal />;
-};
-
-Show.getLayout = function getLayout(page: ReactElement) {
-	return (
-		<Layout.WithSideMenu>
-			<Layout.WithMainMenu showLogo={false}>{page}</Layout.WithMainMenu>
-		</Layout.WithSideMenu>
-	);
-};
-
-export async function getStaticPaths() {
-	return {
-		paths: await getMarketDealsPaths(),
-		fallback: true,
-	};
-}
-
-export async function getStaticProps({ params }) {
-	return { props: params };
-}
-
-export default Show;
