+++ src/pages/[marketplace]/deals/repay.tsx
@@ -12,7 +12,7 @@ import DealAspectGrid from "@components/DealAspectGrid";
 import { Icon } from "@components/Icon";
 import { calculateMonthlyRepaymentAmount } from "@utils/deal.utils";
 import { DealCard } from "@components/DealCard";
-import { getMarketDealsPaths } from "@utils/export.utils";
+import { getMarketsPaths } from "@utils/export.utils";
 
 const Repay: NextPageWithLayout = () => {
 	const router = useRouter();
@@ -133,7 +133,7 @@ Repay.getLayout = function getLayout(page: ReactElement) {
 
 export async function getStaticPaths() {
 	return {
-		paths: await getMarketDealsPaths(),
+		paths: await getMarketsPaths(),
 		fallback: true,
 	};
 }
